Transform codebase into a 💎
