+++ .travis.yml
@@ -1,0 +1,6 @@
+---
+language: ruby
+cache: bundler
+rvm:
+  - 2.6.4
+before_install: gem install bundler -v 2.1.4
