+++ CHANGELOG.md
@@ -1,0 +1,4 @@
+# Changelog
+
+## 0.1.0 10/01/2020
+* Initial gem release
