+++ Gemfile
@@ -1,5 +1,4 @@
 source "https://rubygems.org"
 
-gem "prawn"
-gem "prawn-table"
-gem "thor"
+gemspec
+
