+++ lib/invoice_creator.rb
@@ -1,0 +1,8 @@
+require "invoice_creator/version"
+require "invoice_creator/config_reader"
+require "invoice_creator/models/invoice"
+require "invoice_creator/services/invoice_printer"
+require "invoice_creator/presenters/invoice"
+require "invoice_creator/cli"
+
+module InvoiceCreator; end
