+++ lib/invoice_creator/version.rb
@@ -1,0 +1,3 @@
+module InvoiceCreator
+  VERSION = "0.1.0"
+end
