+++ spec/invoice_creator_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe InvoiceCreator do
+  it "has a version number" do
+    expect(InvoiceCreator::VERSION).not_to be nil
+  end
+end
